requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-ckeditor==6.0.0
django-comments-dab==2.0.0
django-js-asset==1.2.2
django-paypal==1.0.0
gunicorn==20.0.4
idna==2.10
Pillow==8.0.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
